Moved util functions
